move scriptable object to core.editor and make public
