+++ workers/unity/Packages/com.improbable.gdk.buildsystem/Configuration/BuildConfig.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using Improbable.Gdk.Core.Editor;
 using UnityEditor;
 using UnityEngine;
