+++ workers/unity/Packages/com.improbable.gdk.core/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6a92d88d9c7c74b46af1a0a2a8c14424
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
