+++ workers/unity/Packages/com.improbable.gdk.core/Editor/Improbable.Gdk.Core.Editor.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f07b3d1f38b19b1499e5bf3fcadff992
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
