+++ workers/unity/Packages/com.improbable.gdk.core/Editor/SingletonScriptableObject.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e219576071d474c76ba95687b967f8d3
+guid: 692003e7ea20c9c4f90e471e1682b4f0
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
