+++ workers/unity/Packages/com.improbable.gdk.core/Editor/SingletonScriptableObjectLoader.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 730b4dc0c530244c98241974caea5670
+guid: 8d0e7c29342843e43a519703c7081217
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
